refactor(audio): split inference/audio.rs into feature modules
Mechanical split of the ~2.8k-line audio module into decode, telephony,
opus, resample, and pcm submodules under inference/audio/, with stable
re-exports so call sites keep resolving as inference::audio::…. Soft-EOF
Ogg/Opus behavior and all unit/proptests preserved.
